Builds again after stripping many preprocessor statements
