+++ OpenVR-SpaceCalibrator/IPCClient.cpp
@@ -1,10 +1,10 @@
-#include "stdafx.h"
 #include "IPCClient.h"
 #ifdef __linux__
 #include "Comms.h"
 #endif
 
 #include <string>
+#include <iostream>
 
 #ifdef __linux__
 // NOP
