+++ OpenVR-SpaceCalibrator/OpenVR-SpaceCalibrator.h
@@ -1,0 +1,6 @@
+#pragma once
+
+typedef void (*InitErrorCallback)(char const * );
+
+int MainApplication(char const * const cwd, InitErrorCallback cb);
+
