+++ OpenVR-SpaceCalibrator/UserInterface.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include "UserInterface.h"
 #include "Calibration.h"
 #include "Configuration.h"
